cleanup template starter, add ref to create starter
