--- packages/template-blank/.gitignore
@@ -1,34 +1,0 @@
-# --------------------------------
-# Generated by ReNative
-# --------------------------------
-
-.DS_Store
-node_modules/
-npm-debug.log
-yarn-error.log
-.env
-.next
-.expo
-
-#-------------------------------
-# Visual Studio
-#-------------------------------
-.vscode/*
-!.vscode/settings.json
-!.vscode/tasks.json
-!.vscode/launch.json
-!.vscode/extensions.json
-
-#-------------------------------
-# RENATIVE
-#-------------------------------
-
-platformBuilds/
-platformAssets/
-buildHooks/dist
-renative.local.json
-renative.private.json
-renative.build.json
-renative.runtime.json
-rnv-config.local.json
-metro.config.local.js
