--- packages/template-blank/.npmignore
@@ -1,4 +1,0 @@
-buildHooks/dist
-.env
-.npmrc
-docs/
