--- packages/template-blank/README.md
@@ -1,25 +1,0 @@
-# @rnv/template-blank
-
-Blank Template for  <a href="https://www.npmjs.com/package/renative">ReNative</a>
-
-This provides the most basic multi platform setup with only the most fundamental dependencies/plugins added
-
-## 🚀 Quick Start
-
-```bash
-rnv new
-
-...
-
-? What template to use? => '@rnv/template-starter'
-
-```
-
-## Supported Platforms
-
-`iOS`, `tvOS`, `Android`, `Android TV`, `FireTV`, `Android Wear`, `Web`, `Tizen TV`, `Tizen Watch`, `LG webOS`, `macOS/OSX`, `KaiOS`, `FirefoxOS`, `Firefox TV`
-
-
-## Screenshots
-
-<img src="https://github.com/renative-org/renative/blob/develop/packages/template-blank/docs/web.png" width="600px" />
