--- packages/template-blank/appConfigs/blank/renative.json
@@ -1,252 +1,0 @@
-{
-    "id": "blank",
-    "common": {
-        "id": "renative.blank",
-        "title": "ReNative Blank",
-        "description": "Cross-platform application project based on ReNative",
-        "license": "MIT",
-        "includedPlugins": ["*"],
-        "includedFonts": ["*"],
-        "includedPermissions": ["INTERNET"],
-        "backgroundColor": "#FFFFFF"
-    },
-    "platforms": {
-        "ios": {
-            "entryFile": "index",
-            "permissions": ["*"],
-            "buildSchemes": {
-                "debug": {
-                    "runScheme": "Debug",
-                    "bundleAssets": false
-                },
-                "release": {
-                    "runScheme": "Release",
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-        "tvos": {
-            "entryFile": "index",
-            "buildSchemes": {
-                "debug": {
-                    "runScheme": "Debug",
-                    "bundleAssets": false
-                },
-                "release": {
-                    "runScheme": "Release",
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-         "macos": {
-            "entryFile": "index",
-            "provisioningStyle": "Automatic",
-            "permissions": ["*"],
-            "buildSchemes": {
-                "debug": {
-                    "runScheme": "Debug",
-                    "bundleAssets": false,
-                    "environment": "development"
-                },
-                "release": {
-                    "runScheme": "Release",
-                    "bundleAssets": true,
-                    "bundleIsDev": false,
-                    "environment": "production"
-                }
-            }
-        },
-        "android": {
-            "entryFile": "index",
-            "reactNativeEngine": "hermes",
-            "permissions": ["*"],
-            "buildSchemes": {
-                "debug": {
-                    "signingConfig": "Debug",
-                    "bundleAssets": false
-                },
-                "release": {
-                    "signingConfig": "Release",
-                    "bundleAssets": true,
-                    "bundleIsDev": true
-                }
-            }
-        },
-        "androidtv": {
-            "entryFile": "index",
-            "reactNativeEngine": "hermes",
-            "permissions": ["*"],
-            "buildSchemes": {
-                "debug": {
-                    "signingConfig": "Debug",
-                    "bundleAssets": false
-                },
-                "release": {
-                    "signingConfig": "Release",
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-        "firetv": {
-            "entryFile": "index",
-            "reactNativeEngine": "hermes",
-            "permissions": ["*"],
-            "buildSchemes": {
-                "debug": {
-                    "signingConfig": "Debug",
-                    "bundleAssets": false
-                },
-                "release": {
-                    "signingConfig": "Release",
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-        "androidwear": {
-            "entryFile": "index",
-            "reactNativeEngine": "hermes",
-            "permissions": ["*"],
-            "buildSchemes": {
-                "debug": {
-                    "signingConfig": "Debug",
-                    "bundleAssets": false
-                },
-                "release": {
-                    "signingConfig": "Release",
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-        "tizen": {
-            "appName": "ReNativeTV",
-            "package": "NkVRhWHJST",
-            "id": "NkVRhWHJST.ReNativeTV",
-            "entryFile": "index",
-            "certificateProfile": "RNVanillaCert"
-        },
-        "tizenwatch": {
-            "appName": "ReNativeWatch",
-            "package": "cHIP2fIRQZ",
-            "id": "cHIP2fIRQZ.ReNativeWatch",
-            "entryFile": "index",
-            "certificateProfile": "RNVanillaCert"
-        },
-        "tizenmobile": {
-            "appName": "RenativeMobile",
-            "package": "PauodvCU2r",
-            "id": "PauodvCU2r.ReNativeMobile",
-            "entryFile": "index",
-            "certificateProfile": "RNVanillaCert"
-        },
-        "web": {
-            "entryFile": "index",
-            "buildSchemes": {
-                "debug": {
-                    "bundleAssets": false,
-                    "environment": "development"
-                },
-                "release": {
-                    "bundleAssets": true,
-                    "environment": "production"
-                }
-            }
-        },
-        "webtv": {
-            "entryFile": "index",
-            "buildSchemes": {
-                "debug": {
-                    "bundleAssets": false,
-                    "environment": "development"
-                },
-                "release": {
-                    "bundleAssets": true,
-                    "environment": "production"
-                }
-            }
-        },
-        "webos": {
-            "entryFile": "index"
-        },
-        "windows": {
-            "entryFile": "index",
-            "language": "cpp",
-            "buildSchemes": {
-                "debug": {
-                    "bundleAssets": false
-                },
-                "release": {
-                    "release": true,
-                    "bundleAssets": true,
-                    "bundleIsDev": false,
-                    "enableSourceMaps": true
-                }
-            }
-        },
-        "xbox": {
-            "entryFile": "index",
-            "language": "cpp",
-            "arch": "x64",
-            "buildSchemes": {
-                "debug": {
-                    "bundleAssets": false
-                },
-                "release": {
-                    "release": true,
-                    "bundleAssets": true,
-                    "bundleIsDev": false,
-                    "enableSourceMaps": true
-                }
-            }
-        },
-        "kaios": {
-            "entryFile": "index",
-            "buildSchemes": {
-                "debug": {
-                    "bundleAssets": false
-                },
-                "release": {
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-        "firefoxos": {
-            "entryFile": "index",
-            "buildSchemes": {
-                "debug": {
-                    "bundleAssets": false
-                },
-                "release": {
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-        "firefoxtv": {
-            "entryFile": "index",
-            "buildSchemes": {
-                "debug": {
-                    "bundleAssets": false
-                },
-                "release": {
-                    "bundleAssets": true,
-                    "bundleIsDev": false
-                }
-            }
-        },
-        "chromecast": {
-            "entryFile": "index",
-            "webpackConfig": {
-                "customScripts": [
-                    "www.gstatic.com/cast/sdk/libs/caf_receiver/v3/cast_receiver_framework.js",
-                    "www.gstatic.com/cast/sdk/libs/devtools/debug_layer/caf_receiver_logger.js"
-                ]
-            }
-        }
-    }
-}
