--- packages/template-blank/babel.config.js
@@ -1,3 +1,0 @@
-const { withRNVBabel } = require('rnv');
-
-module.exports = withRNVBabel({});
