--- packages/template-blank/buildHooks/src/index.js
@@ -1,11 +1,0 @@
-import { Logger } from 'rnv';
-
-const hooks = {
-    hello: async () => {
-        Logger.logHook('Hello build hook!');
-    }
-};
-
-const pipes = {};
-
-export { pipes, hooks };
