--- packages/template-blank/index.js
@@ -1,1 +1,0 @@
-import './src/entry';// eslint-disable-line
