--- packages/template-blank/metro.config.js
@@ -1,5 +1,0 @@
-const { withRNVMetro } = require('rnv');
-
-const defaultConfig = {};
-
-module.exports = withRNVMetro(defaultConfig);
