--- packages/template-blank/next.config.js
@@ -1,5 +1,0 @@
-const { withRNVMetro } = require('rnv');
-
-module.exports = withRNVMetro({
-    compress: false
-});
