--- packages/template-blank/package.json
@@ -1,85 +1,0 @@
-{
-    "name": "@rnv/template-blank",
-    "version": "0.37.0-rc.4",
-    "description": "🚧 Blank Template for ReNative (https://www.npmjs.com/package/renative). Supports `iOS`, `tvOS`, `Android`, `Android TV`, `FireTV`, `Android Wear`, `Web`, `Tizen TV`, `Tizen Watch`, `LG webOS`, `macOS/OSX`, `Windows`, `KaiOS`, `FirefoxOS`, `Firefox TV`",
-    "keywords": [
-        "android tv",
-        "android wear",
-        "apple tv",
-        "boilerplate",
-        "cocoapods",
-        "cross-platform",
-        "desktop",
-        "electron",
-        "firefox tv",
-        "firefoxox",
-        "kaios",
-        "kit",
-        "kotlin",
-        "lg",
-        "macos",
-        "multiplatform",
-        "native",
-        "osx",
-        "react",
-        "react-native",
-        "smart tv",
-        "starter",
-        "swift",
-        "template",
-        "tizen",
-        "tizen watch",
-        "tvos",
-        "watch",
-        "web",
-        "webos",
-        "windows"
-    ],
-    "homepage": "https://github.com/renative-org/renative#readme",
-    "bugs": {
-        "url": "https://github.com/renative-org/renative/issues"
-    },
-    "license": "MIT",
-    "author": "Pavel Jacko <pavel.jacko@gmail.com> (https://github.com/pavjacko)",
-    "contributors": [],
-    "files": [
-        "appConfigs",
-        "buildHooks",
-        "index.js",
-        "next.config.js",
-        "renative.json",
-        "renative.template.json",
-        "src",
-        "static",
-        "tsconfig.json"
-    ],
-    "repository": {
-        "type": "git",
-        "url": "git://github.com/renative-org/renative.git"
-    },
-    "scripts": {
-        "tsc": "tsc --noEmit --composite false"
-    },
-    "devDependencies": {
-        "@rnv/engine-rn": "0.37.0-rc.4",
-        "@rnv/engine-rn-electron": "0.37.0-rc.4",
-        "@rnv/engine-rn-next": "0.37.0-rc.4",
-        "@rnv/engine-rn-tvos": "0.37.0-rc.4",
-        "@rnv/engine-rn-web": "0.37.0-rc.4",
-        "@rnv/engine-rn-windows": "0.37.0-rc.4",
-        "@rnv/renative": "0.37.0-rc.4",
-        "@types/react": "18.0.11",
-        "@types/react-dom": "18.0.5",
-        "@types/react-native": "0.67.2",
-        "react": "18.1.0",
-        "react-art": "18.1.0",
-        "react-dom": "18.1.0",
-        "react-native": "0.67.2",
-        "rnv": "0.37.0-rc.4"
-    },
-    "private": false,
-    "publishConfig": {
-        "access": "public"
-    },
-    "gitHead": "930edaf234cc321ad9f1e296c5d394e3144845e2"
-}
