--- packages/template-blank/renative.json
@@ -1,201 +1,0 @@
-{
-    "engines": {
-        "@rnv/engine-lightning": "source:rnv",
-        "@rnv/engine-rn": "source:rnv",
-        "@rnv/engine-rn-electron": "source:rnv",
-        "@rnv/engine-rn-macos": "source:rnv",
-        "@rnv/engine-rn-next": "source:rnv",
-        "@rnv/engine-rn-tvos": "source:rnv",
-        "@rnv/engine-rn-web": "source:rnv",
-        "@rnv/engine-rn-windows": "source:rnv"
-    },
-    "isTemplate": true,
-    "projectName": "renative",
-    "enableHookRebuild": false,
-    "workspaceID": "rnv",
-    "isMonorepo": true,
-    "paths": {
-        "appConfigsDir": "./appConfigs",
-        "platformAssetsDir": "./platformAssets",
-        "platformBuildsDir": "./platformBuilds"
-    },
-    "tasks": {
-        "install": {
-            "script": "yarn bootstrap"
-        }
-    },
-    "defaults": {
-        "ports": {
-            "ios": 8182,
-            "android": 8183,
-            "androidtv": 8184,
-            "firetv": 8198,
-            "tvos": 8185,
-            "macos": 8186,
-            "web": 8180,
-            "tizen": 8187,
-            "webos": 8188,
-            "androidwear": 8189,
-            "tizenwatch": 8190,
-            "tizenmobile": 8191,
-            "windows": 8192,
-            "xbox": 8199,
-            "kaios": 8193,
-            "firefoxos": 8194,
-            "firefoxtv": 8114
-        },
-        "supportedPlatforms": [
-            "ios",
-            "android",
-            "androidtv",
-            "firetv",
-            "androidwear",
-            "web",
-            "webtv",
-            "tizen",
-            "tvos",
-            "webos",
-            "macos",
-            "windows",
-            "tizenwatch",
-            "tizenmobile",
-            "kaios",
-            "firefoxos",
-            "firefoxtv",
-            "chromecast",
-            "xbox"
-        ],
-        "schemes": {
-            "run": "debug",
-            "build": "release",
-            "export": "release"
-        },
-        "targets": {}
-    },
-    "templates": {
-        "@rnv/template-starter": {
-            "version": "^0.37.0-rc.4",
-            "webpack": {
-                "moduleAliases": true
-            }
-        }
-    },
-    "platforms": {
-        "android": {
-            "gradle.properties": {
-                "android.useDeprecatedNdk": true
-            },
-            "AndroidManifest": {
-                "children": [
-                    {
-                        "tag": "application",
-                        "android:name": ".MainApplication",
-                        "android:usesCleartextTraffic": true,
-                        "tools:targetApi": 28
-                    }
-                ]
-            },
-            "mainActivity": {
-                "onCreate": "super.onCreate(null)"
-            },
-            "includedPermissions": [
-                "INTERNET"
-            ]
-        },
-        "androidtv": {
-            "gradle.properties": {
-                "android.useDeprecatedNdk": true
-            },
-            "AndroidManifest": {
-                "children": [
-                    {
-                        "tag": "application",
-                        "android:name": ".MainApplication",
-                        "android:usesCleartextTraffic": true,
-                        "tools:targetApi": 28
-                    }
-                ]
-            },
-            "engine": "engine-rn-tvos",
-            "includedPermissions": [
-                "INTERNET"
-            ]
-        },
-        "androidwear": {
-            "gradle.properties": {
-                "android.useDeprecatedNdk": true
-            },
-            "AndroidManifest": {
-                "children": [
-                    {
-                        "tag": "application",
-                        "android:name": ".MainApplication",
-                        "android:usesCleartextTraffic": true,
-                        "tools:targetApi": 28
-                    }
-                ]
-            },
-            "includedPermissions": [
-                "INTERNET"
-            ],
-            "mainActivity": {
-                "onCreate": "super.onCreate(null)"
-            }
-        },
-        "firetv": {
-            "gradle.properties": {
-                "android.useDeprecatedNdk": true
-            },
-            "AndroidManifest": {
-                "children": [
-                    {
-                        "tag": "application",
-                        "android:name": ".MainApplication",
-                        "android:usesCleartextTraffic": true,
-                        "tools:targetApi": 28
-                    }
-                ]
-            },
-            "engine": "engine-rn-tvos",
-            "includedPermissions": [
-                "INTERNET"
-            ]
-        },
-        "web": {
-            "engine": "engine-rn-next"
-        },
-        "tvos": {
-            "engine": "engine-rn-tvos"
-        },
-        "macos": {
-            "engine": "engine-rn"
-        },
-        "windows": {
-            "engine": "engine-rn-electron"
-        },
-        "xbox": {
-            "engine": "engine-rn-windows"
-        },
-        "ios": {
-            "deploymentTarget": "14.0"
-        }
-    },
-    "plugins": {
-        "react-native-web-image-loader": "source:rnv",
-        "hash-source": "1.0.4",
-        "@rnv/renative": "source:rnv"
-    },
-    "permissions": {
-        "ios": {},
-        "android": {
-            "INTERNET": {
-                "key": "android.permission.INTERNET",
-                "security": "normal"
-            },
-            "SYSTEM_ALERT_WINDOW": {
-                "key": "android.permission.SYSTEM_ALERT_WINDOW",
-                "security": "signature"
-            }
-        }
-    }
-}
