--- packages/template-blank/renative.template.json
@@ -1,33 +1,0 @@
-{
-    "templateConfig": {
-        "includedPaths": [
-            "buildHooks",
-            "src",
-            "index.js",
-            "tsconfig.json"
-        ],
-        "packageTemplate": {
-            "devDependencies": {
-                "typescript": "4.5.4",
-                "@types/react": "18.0.11",
-                "@types/react-dom": "18.0.5",
-                "@types/react-native": "0.67.2"
-            },
-            "browserslist": {
-                "production": [
-                    ">0.2%",
-                    "not dead",
-                    "not op_mini all"
-                ],
-                "development": [
-                    "last 1 chrome version",
-                    "last 1 firefox version",
-                    "last 1 safari version"
-                ]
-            }
-        }
-    },
-    "platforms": {},
-    "plugins": {},
-    "extendsTemplate": "@rnv/template-starter/renative.json"
-}
