--- packages/template-blank/src/entry/index.ts
@@ -1,5 +1,0 @@
-import "react-native/Libraries/Core/InitializeCore"; //https://github.com/renative-org/renative/issues/876
-import { AppRegistry } from 'react-native';
-import App from '../app';
-
-AppRegistry.registerComponent('App', () => App);
