--- packages/template-blank/tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "extends": "@flexn/typescript/tsconfig.app.json",
-    "compilerOptions": {
-        "outDir": "./lib",
-        "rootDir": "."
-    },
-    "include": [
-        "./src/**/*",
-        "./src/**/*.json",
-        "./types.d.ts",
-        "./package.json",
-        "./app.json",
-        "./platformAssets/runtime/fontManager.js",
-        "./platformAssets/runtime/logo.png"
-    ]
-}
