+++ packages/template-starter/package.json
@@ -76,71 +76,39 @@
         "tsc": "tsc --noEmit --composite false"
     },
     "devDependencies": {
-        "@babel/runtime": "7.12.5",
         "@flexn/graybox": "0.21.2-alpha.16",
         "@lightningjs/cli": "2.8.0",
         "@lightningjs/sdk": "4.8.3",
-        "@noriginmedia/react-spatial-navigation": "2.12.9",
         "@react-native-community/cli": "^6.0.0",
         "@react-native-community/cli-platform-android": "^6.0.0",
         "@react-native-community/cli-platform-ios": "^6.0.0",
-        "@react-native-community/masked-view": "0.1.6",
-        "@react-native-community/viewpager": "2.0.2",
         "@react-native-windows/cli": "0.67.1",
-        "@react-navigation/bottom-tabs": "5.1.1",
-        "@react-navigation/core": "5.12.3",
-        "@react-navigation/drawer": "5.9.0",
-        "@react-navigation/material-bottom-tabs": "5.1.1",
-        "@react-navigation/material-top-tabs": "5.1.1",
-        "@react-navigation/native": "5.7.3",
-        "@react-navigation/routers": "5.1.0",
-        "@react-navigation/stack": "5.9.0",
-        "@react-navigation/web": "1.0.0-alpha.9",
         "@rnv/engine-lightning": "0.37.0-rc.4",
         "@rnv/engine-rn": "0.37.0-rc.4",
         "@rnv/engine-rn-electron": "0.37.0-rc.4",
-        "@rnv/engine-rn-macos": "0.37.0-rc.4",
         "@rnv/engine-rn-next": "0.37.0-rc.4",
         "@rnv/engine-rn-tvos": "0.37.0-rc.4",
         "@rnv/engine-rn-web": "0.37.0-rc.4",
         "@rnv/engine-rn-windows": "0.37.0-rc.4",
         "@rnv/renative": "0.37.0-rc.4",
-        "@rnv/template-starter": "0.37.0-rc.4",
         "@types/react": "18.0.11",
         "@types/react-dom": "18.0.5",
         "@types/react-native": "0.67.2",
         "babel-jest": "26.6.3",
         "babel-loader": "8.0.5",
-        "cypress": "4.8.0",
         "detox": "18.20.2",
-        "hash-source": "1.0.4",
         "jetifier": "1.6.5",
         "next": "12.1.6",
         "raf": "3.4.1",
-        "react": "18.1.0",
-        "react-art": "18.1.0",
-        "react-dom": "18.1.0",
+        "react": "18.2.0",
+        "react-art": "18.2.0",
+        "react-dom": "18.2.0",
         "react-native": "0.67.2",
-        "react-native-gesture-handler": "1.10.3",
-        "react-native-google-cast": "4.2.4",
-        "react-native-macos": "0.66.33",
-        "react-native-reanimated": "1.13.3",
-        "react-native-safe-area-context": "3.1.0",
-        "react-native-safe-area-view": "0.14.5",
-        "react-native-screens": "2.2.0",
-        "react-native-tab-view": "2.13.0",
         "react-native-tvos": "0.66.3-1",
-        "react-native-vector-icons": "8.1.0",
         "react-native-web": "0.18.0",
-        "react-native-web-image-loader": "0.0.5",
         "react-native-windows": "0.69.4",
-        "react-router-dom": "6.3.0",
         "rnv": "0.37.0-rc.4",
-        "v8-android": "10.100.1",
         "xcode": "2.1.0"
-    },
-    "optionalDependencies": {
-        "ios-deploy": "1.11.4"
     },
     "private": false,
     "publishConfig": {
