--- packages/template-starter/src/components/CastButton/index.desktop.web.ts
@@ -1,3 +1,0 @@
-export const CastButton = () => {
-    return null; //not supported in web
-};
