--- packages/template-starter/src/components/CastButton/index.tsx
@@ -1,18 +1,0 @@
-// import { CastButton } from 'react-native-google-cast';
-import React from 'react';
-import { View } from "react-native";
-
-// export { CastButton };
-
-// To enable it add casting add this to renative.json plugins section:
-// "react-native-google-cast": {
-//     "props": {
-//         "applicationID": "CC1AD845"
-//     }
-// },
-// ...and return CastButton from 'react-native-google-cast
-
-
-export const CastButton = () => {
-    return <View />
-};
