+++ packages/template-starter/src/entry/index.lng.ts
