+++ packages/template-starter/src/entry/index.web.ts
