--- packages/template-starter/src/hooks/linking.ts
@@ -1,24 +1,0 @@
-import { Linking } from 'react-native';
-import {
-    isPlatformIos,
-    isPlatformAndroid,
-    isPlatformMacos,
-    isPlatformWeb
-} from '@rnv/renative';
-
-export function useOpenURL() {
-    function openURL(url) {
-        if (
-            isPlatformIos
-            || isPlatformAndroid
-            || isPlatformMacos
-            || isPlatformWeb
-        ) {
-            // eslint-disable-next-line no-console
-            Linking.openURL(url).catch(err => console.error('An error occurred', err));
-        } else {
-            // Not supported
-        }
-    }
-    return openURL;
-}
