--- packages/template-starter/src/navigation/index.chromecast.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import ScreenCast from '../screens/Cast';
-
-
-const App = () => (
-    <ScreenCast />
-);
-
-export default App;
