--- packages/template-starter/src/utils/index.ts
@@ -1,17 +1,0 @@
-import {
-    isPlatformIos,
-    isPlatformMacos,
-    isPlatformTvos,
-    isPlatformWeb,
-} from '@rnv/renative';
-
-export function testProps(testId: string | undefined) {
-    if (!testId) {
-        return;
-    }
-    const isApplePlatform = isPlatformIos || isPlatformTvos || isPlatformMacos;
-    if (isApplePlatform || isPlatformWeb) {
-        return { testID: testId };
-    }
-    return { accessibilityLabel: testId, accessible: true };
-}
